chore: moved mutations into hooks
