+++ src/renderer/src/hooks/mutations/deviceInfo.ts
