+++ src/renderer/src/routes/Onboarding/CreateProjectScreen.tsx
@@ -20,8 +20,8 @@ import {
 } from '../../components/Onboarding/onboardingLogic'
 import { Text } from '../../components/Text'
 import { PROJECT_NAME_MAX_LENGTH_GRAPHEMES } from '../../constants'
+import { useCreateProject } from '../../hooks/mutations/projects'
 import ProjectImage from '../../images/add_square.png'
-import { useCreateProject } from '../../mutations/projects'
 
 export const m = defineMessages({
 	title: {
